// ui/src/components/__generated__/BookmarkListItemFragment.ts
/* tslint:disable */
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL fragment: BookmarkListItemFragment
// ====================================================

export interface BookmarkListItemFragment_user {
  name: string;
}

export interface BookmarkListItemFragment_entry {
  id: string;
  title: string;
  url: string;
}

export interface BookmarkListItemFragment {
  user: BookmarkListItemFragment_user;
  entry: BookmarkListItemFragment_entry;
  comment: string;
}

/* tslint:disable */
// This file was automatically generated and should not be edited.

//==============================================================
// START Enums and Input Objects
//==============================================================

//==============================================================
// END Enums and Input Objects
//==============================================================